Clean up build warnings C4018 signed/unsigned mismatch in '<'.

+++ tools/optimizer/parser.h
@@ -687,11 +687,11 @@ class Parser {
             if (parts[i].isNode) continue;
             IString op = parts[i].getOp();
             if (!ops.ops.has(op)) continue;
-            if (ops.type == OperatorClass::Binary && i > 0 && i < parts.size()-1) {
+            if (ops.type == OperatorClass::Binary && i > 0 && i < (int)parts.size()-1) {
               parts[i] = Builder::makeBinary(parts[i-1].getNode(), op, parts[i+1].getNode());
               parts.erase(parts.begin() + i + 1);
               parts.erase(parts.begin() + i - 1);
-            } else if (ops.type == OperatorClass::Prefix && i < parts.size()-1) {
+            } else if (ops.type == OperatorClass::Prefix && i < (int)parts.size()-1) {
               if (i > 0 && parts[i-1].isNode) continue; // cannot apply prefix operator if it would join two nodes
               parts[i] = Builder::makePrefix(op, parts[i+1].getNode());
               parts.erase(parts.begin() + i + 1);
@@ -700,7 +700,7 @@ class Parser {
               //                      ^
               //dumpParts(parts, i);
               if (op != COLON) continue;
-              assert(i < parts.size()-1 && i >= 3);
+              assert(i < (int)parts.size()-1 && i >= 3);
               if (parts[i-2].getOp() != QUESTION) continue; // e.g. x ? y ? 1 : 0 : 2
               parts[i-3] = Builder::makeConditional(parts[i-3].getNode(), parts[i-1].getNode(), parts[i+1].getNode());
               parts.erase(parts.begin() + i - 2, parts.begin() + i + 2);
@@ -709,16 +709,16 @@ class Parser {
           }
         } else {
           // left to right
-          for (int i = 0; i < parts.size(); i++) {
+          for (int i = 0; i < (int)parts.size(); i++) {
             if (parts[i].isNode) continue;
             IString op = parts[i].getOp();
             if (!ops.ops.has(op)) continue;
-            if (ops.type == OperatorClass::Binary && i > 0 && i < parts.size()-1) {
+            if (ops.type == OperatorClass::Binary && i > 0 && i < (int)parts.size()-1) {
               parts[i] = Builder::makeBinary(parts[i-1].getNode(), op, parts[i+1].getNode());
               parts.erase(parts.begin() + i + 1);
               parts.erase(parts.begin() + i - 1);
               i--;
-            } else if (ops.type == OperatorClass::Prefix && i < parts.size()-1) {
+            } else if (ops.type == OperatorClass::Prefix && i < (int)parts.size()-1) {
               if (i > 0 && parts[i-1].isNode) continue; // cannot apply prefix operator if it would join two nodes
               parts[i] = Builder::makePrefix(op, parts[i+1].getNode());
               parts.erase(parts.begin() + i + 1);
